fix(#336): Cross-Tab-Saldo in Drill-Log + Maschinen-Protokoll, NICHT im Ergebnis-Tab
User-Feedback 2026-06-23 (5. Runde): „Es soll in den Ergebnissen
Bereich unter dünger verbleibend im Protokoll/drill log".

Korrektur zu PR #343: ich hatte den Saldo in den Ergebnis-Tab gesetzt,
der User meinte aber den Drill-Log + Maschinen-Protokoll. „Dünger
verbleibend" steht im Drill-Log (renderDrillSummary), nicht im
Ergebnis-Tab. Mein Fehler.

Layout jetzt:

- Ergebnis-Tab (gelb): Per-Tab-Shape zurück (PR #337/#339):
    Label: „Abweichung dieses Tabs"
    Zeile: „Ersparnis: X Einheiten Saatgut, Y kg Dünger" (per aktiver Tab)
    KEIN .net-totals-*
- Drill-Log + Maschinen-Protokoll: Cross-Tab-Saldo als erster Block
  VOR den Per-Tab-Headern. Label „Gesamt-Saldo (alle Tabs)" + grüne/
  rote Saldo-Zeile. Darunter die Per-Tab-Header + drill-savings/
  drill-carryover/drill-excess Blöcke wie bisher.

Helper _appendNetTotalsBlock ist zurück (war in #343 entfernt, jetzt
wieder da). _computeTabSelfSaldo bleibt (Pattern 3 Single Source of
Truth).

Tests: tests/50-/51- reaktiviert (Per-Tab-Shape im Ergebnis-Tab ist
zurück). tests/52 umbenannt zu drill-protokoll-saldo und
umgeschrieben: 12 Tests für Saldo in Drill-Log + Maschinen-Protokoll
(Worked Example 1,0/100, Reihenfolge, Multi-Tab, Pure Mehrbedarf,
Re-Render Hygiene, Per-Tab drill-savings bleibt).

pnpm vitest run: 48/48 files, 828/828 grün.
pnpm lint: clean (kein eslint-disable nötig diesmal, der
Net-Totals-Helper ist wieder nur in render-drill.js definiert).

CACHE_VERSION v31 → v32: invalidiert SW-Cache.
